requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=1.30.0
python-docx>=1.1.0
vllm
